+++ packages/core/src/hid/index.ts
@@ -1,2 +1,3 @@
+export * from "./hid-buttons";
 export * from "./hid-gamepad";
 export * from "./xinput";